fix: resolve type mismatches in LEM recording modules
- Fix fov field type (float -> u32) in metadata_writer.rs
- Fix return type in tokio_thread.rs tick function
- Fix vkey type cast (u16 -> u32) in lem_input_recorder.rs
- Add MESSAGEBOX_STYLE import and wrap in notification.rs

+++ src/record/lem_input_recorder.rs
@@ -278,7 +278,7 @@ fn vkey_to_string(vkey: u16) -> String {
         0x26 => "Up".to_string(),
         0x27 => "Right".to_string(),
         0x28 => "Down".to_string(),
-        0x30..=0x39 => char::from_digit(vkey - 0x30, 10).unwrap().to_string(),
+        0x30..=0x39 => char::from_digit((vkey - 0x30) as u32, 10).unwrap().to_string(),
         0x41..=0x5A => char::from_u32(vkey as u32 - 0x41 + b'A' as u32)
             .unwrap()
             .to_string(),
